NOJIRA: Bruker fiksen som kom i TFP-4351 på SVP innvilgelse også, slik at vi slipper å ha koden duplisert. Fjernet bruk av utgåtte DokumentMalTyper for brev som ble lansert på Fpdokgen for lenge siden.
