+++ brevproduksjon/src/main/java/no/nav/foreldrepenger/fpformidling/brevproduksjon/mapper/innvilgelsefp/BeregningsgrunnlagMapper.java
@@ -1,13 +1,12 @@
 package no.nav.foreldrepenger.fpformidling.brevproduksjon.mapper.innvilgelsefp;
 
+import static no.nav.foreldrepenger.fpformidling.brevproduksjon.mapper.felles.BeregningsgrunnlagMapper.finnAktivitetStatuserForAndeler;
+
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 import no.nav.foreldrepenger.fpformidling.beregningsgrunnlag.AktivitetStatus;
 import no.nav.foreldrepenger.fpformidling.beregningsgrunnlag.BGAndelArbeidsforhold;
@@ -22,18 +21,6 @@ import no.nav.foreldrepenger.fpformidling.opptjening.OpptjeningAktivitetType;
 import no.nav.foreldrepenger.fpformidling.virksomhet.Arbeidsgiver;
 
 public final class BeregningsgrunnlagMapper {
-
-    private static final Map<AktivitetStatus, List<AktivitetStatus>> kombinerteRegelStatuserMap = new HashMap<>();
-
-    static {
-        kombinerteRegelStatuserMap.put(AktivitetStatus.KOMBINERT_AT_FL, List.of(AktivitetStatus.ARBEIDSTAKER, AktivitetStatus.FRILANSER));
-        kombinerteRegelStatuserMap.put(AktivitetStatus.KOMBINERT_AT_SN,
-                List.of(AktivitetStatus.ARBEIDSTAKER, AktivitetStatus.SELVSTENDIG_NÆRINGSDRIVENDE));
-        kombinerteRegelStatuserMap.put(AktivitetStatus.KOMBINERT_AT_FL_SN,
-                List.of(AktivitetStatus.ARBEIDSTAKER, AktivitetStatus.FRILANSER, AktivitetStatus.SELVSTENDIG_NÆRINGSDRIVENDE));
-        kombinerteRegelStatuserMap.put(AktivitetStatus.KOMBINERT_FL_SN,
-                List.of(AktivitetStatus.FRILANSER, AktivitetStatus.SELVSTENDIG_NÆRINGSDRIVENDE));
-    }
 
     public static Beløp finnBrutto(Beregningsgrunnlag beregningsgrunnlag) {
         double sum = finnBgpsaListe(beregningsgrunnlag).stream()
@@ -83,57 +70,6 @@ public final class BeregningsgrunnlagMapper {
 
     private static List<BeregningsgrunnlagPrStatusOgAndel> finnBgpsaListe(Beregningsgrunnlag beregningsgrunnlag) {
         return finnFørstePeriode(beregningsgrunnlag).getBeregningsgrunnlagPrStatusOgAndelList();
-    }
-
-    private static List<BeregningsgrunnlagPrStatusOgAndel> finnAktivitetStatuserForAndeler(BeregningsgrunnlagAktivitetStatus bgAktivitetStatus,
-            List<BeregningsgrunnlagPrStatusOgAndel> bgpsaListe) {
-        List<BeregningsgrunnlagPrStatusOgAndel> resultatListe;
-        if (AktivitetStatus.KUN_YTELSE.equals(bgAktivitetStatus.aktivitetStatus())) {
-            return bgpsaListe;
-        }
-        if (bgAktivitetStatus.aktivitetStatus().harKombinertStatus()) {
-            List<AktivitetStatus> relevanteStatuser = kombinerteRegelStatuserMap.get(bgAktivitetStatus.aktivitetStatus());
-            resultatListe = bgpsaListe.stream().filter(andel -> relevanteStatuser.contains(andel.getAktivitetStatus())).collect(Collectors.toList());
-        } else {
-            resultatListe = bgpsaListe.stream().filter(andel -> bgAktivitetStatus.aktivitetStatus().equals(andel.getAktivitetStatus()))
-                    .collect(Collectors.toList());
-
-            // Spesialhåndtering av tilkommet arbeidsforhold for Dagpenger og AAP - andeler
-            // som ikke kan mappes gjennom aktivitetesstatuslisten på beregningsgrunnlag da
-            // de er tilkommet etter skjæringstidspunkt. Typisk dersom arbeidsgiver er
-            // tilkommet etter start permisjon og krever refusjon i permisjonstiden.
-            List<AktivitetStatus> aktuelleStatuserForTilkommetArbForhold = List.of(AktivitetStatus.DAGPENGER,
-                    AktivitetStatus.ARBEIDSAVKLARINGSPENGER);
-            if (resultatListe.stream().anyMatch(br -> aktuelleStatuserForTilkommetArbForhold.contains(br.getAktivitetStatus()))
-                    && hentSummertDagsats(resultatListe) != hentSummertDagsats(bgpsaListe)) {
-                long sumTilkommetDagsats = hentSumTilkommetDagsats(bgpsaListe);
-                if (sumTilkommetDagsats != 0) {
-                    resultatListe.forEach(rl -> {
-                        if (aktuelleStatuserForTilkommetArbForhold.contains(rl.getAktivitetStatus())) {
-                            rl.setDagsats(sumTilkommetDagsats);
-                        }
-                    });
-                }
-            }
-        }
-        if (resultatListe.isEmpty()) {
-            StringBuilder sb = new StringBuilder();
-            bgpsaListe.stream().map(BeregningsgrunnlagPrStatusOgAndel::getAktivitetStatus).map(AktivitetStatus::getKode).forEach(sb::append);
-            throw new IllegalStateException(String.format("Fant ingen andeler for status: %s, andeler: %s", bgAktivitetStatus.aktivitetStatus(), sb));
-        }
-        return resultatListe;
-    }
-
-    private static long hentSummertDagsats(List<BeregningsgrunnlagPrStatusOgAndel> bgpsaListe) {
-        return bgpsaListe.stream().map(BeregningsgrunnlagPrStatusOgAndel::getDagsats).reduce(Long::sum).orElse(0L);
-    }
-
-    private static long hentSumTilkommetDagsats(List<BeregningsgrunnlagPrStatusOgAndel> bgpsaListe) {
-        return bgpsaListe.stream()
-                .filter(andel -> andel.getDagsats() > 0)
-                .filter(BeregningsgrunnlagPrStatusOgAndel::getErTilkommetAndel)
-                .map(BeregningsgrunnlagPrStatusOgAndel::getDagsats)
-                .reduce(Long::sum).orElse(0L);
     }
 
     private static BeregningsgrunnlagAndel lagBeregningsgrunnlagAndel(BeregningsgrunnlagPrStatusOgAndel andel) {
